refactor(tui): remove test-utils barrel re-export
