+++ src/chat.tui.test.tsx
@@ -5,7 +5,7 @@ import { ChatInputPanel } from "./chat-input-panel";
 import { palette } from "./palette";
 import { dedent } from "./test-utils";
 import { DEFAULT_TERMINAL_WIDTH } from "./tui/constants";
-import { renderPlain } from "./tui-test-utils";
+import { renderPlain } from "./tui/test-utils";
 
 const DEFAULT_FOOTER_CONTEXT = "~/code/acolyte · main";
