--- src/tui-test-utils.ts
@@ -1,8 +1,0 @@
-export {
-  renderCapture,
-  renderHook,
-  renderPlain,
-  trimRightLines,
-  wait,
-  withTerminalWidth,
-} from "./tui/test-utils";
